refactor and completed Activity views | Elena

+++ app/Http/Controllers/ActivityController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Activity;
+use App\User;
+use App\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ActivityController extends Controller
 {
@@ -16,11 +19,19 @@ class ActivityController extends Controller
     public function index()
     {
         $activities = Activity::all();
-        return view('activity.index', compact('activities'));
+        $users = User::all();
+        $roles = Role::all();
+        return view('activity.index', ['users' => $users], compact('activities'), compact('roles'));
     }
 
     public function create()
     {
+        $activities = Activity::all();
+        $users = User::all();
+        $roles = Role::all();
+        if (auth()->user()->role_id === "Soci") {
+            return view('user.notauthorized');
+        }
         return view('activity.create');
     }
 
@@ -32,13 +43,18 @@ class ActivityController extends Controller
 
     public function show(Activity $activity)
     {
-        //
+        $roles = Role::all();
+        return view('activity.show', ['users' => $users], compact('activities'), compact('roles'));
     }
 
     public function edit(Activity $activity)
     {
-        return view('activity.edit', compact('activity'));
+        $roles = Role::all();
+        if (auth()->user()->can('edit', $user)) {
+            return view('activity.edit', ['users' => $users], compact('activities'), compact('roles'));
+        }
     }
+    
     public function update(Request $request, Activity $activity)
     {
         $activity->update($request->all());
@@ -47,7 +63,9 @@ class ActivityController extends Controller
 
     public function destroy(Activity $activity)
     {
-        $activity->delete();
+        if (auth()->user()->can('destroy', $activity)) {
+            $activity->delete();
+        }
         return redirect('/activity');
     }
 }
